+++ src/main/java/fr/maxlego08/essentials/hologram/HologramModule.java
@@ -147,7 +147,7 @@ public class HologramModule extends ZModule implements HologramManager {
         Hologram hologram = this.createHologram(hologramType, hologramConfiguration, name, name, player.getLocation());
 
         if (hologramType == HologramType.TEXT) {
-            hologram.addLine(new ZHologramLine(1, "&fUse #ff9966/holo editline " + name + " 1 <your text>&f!"));
+            hologram.addLine(new ZHologramLine(1, "&fUse #ff9966/holo setline " + name + " 1 <your text>&f!"));
         }
 
         hologram.create();
@@ -259,7 +259,7 @@ public class HologramModule extends ZModule implements HologramManager {
     }
 
     @EventHandler
-    public void onQuid(PlayerQuitEvent event) {
+    public void onQuit(PlayerQuitEvent event) {
         Player player = event.getPlayer();
         this.holograms.forEach(hologram -> hologram.removePlayer(player));
     }