+++ src/main/java/fr/maxlego08/essentials/storage/database/repositeries/UserRepository.java
@@ -230,4 +230,11 @@ public class UserRepository extends Repository {
             }
         });
     }
+
+    public void updateFrozen(UUID uuid, boolean frozen) {
+        update(table -> {
+            table.bool("frozen", frozen);
+            table.where("unique_id", uuid);
+        });
+    }
 }